refactor, clean up
