+++ data-structures/YASL_Table.c
@@ -7,6 +7,7 @@
 #include "hash_function.h"
 #include "interpreter/refcount.h"
 #include "interpreter/YASL_Object.h"
+#include "interpreter/userdata.h"
 
 struct YASL_Table_Item TOMBSTONE = { { Y_END, { Y_END } }, { Y_END, { Y_END } } };
