+++ util/prime.c
@@ -5,7 +5,7 @@
 
 const unsigned PRIMES_UNDER_200[] = {2,3,5,7,11,13,17,19,23,29,31,37,41,43,47,53,59,61,67,71,73,79,83,89,97,101,103,107,109,113,127,131,137,139,149,151,157,163,167,173,179,181,191,193,197,199};
 
-// returns 1 -- prime, 0 -- not prime or undefined
+// returns 1 -- util, 0 -- not util or undefined
 bool is_prime(const size_t x) {
 	if (x < 2) return false;
 	if (x < 4) return true;
